requirements.txt
influxdb==5.2.3
python-dotenv==0.10.5